draw control componenet
