+++ src/styles/components/_c-zoom-control.scss
@@ -5,10 +5,10 @@
   display: flex;
   flex-direction: column;
   z-index: 1000;
-  
+
   .zoom-control-btn {
     background-color: $theme-base;
-    padding: 5px;
+    padding: 10px 5px;
     margin-bottom: 2px;
     cursor: pointer;
   }
